requirements.txt
asgiref==3.6.0
Django==4.1.6
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2==2.9.5
PyJWT==2.6.0
python-decouple==3.7
pytz==2022.7.1
sqlparse==0.4.3
